+++ ch2-sorting-algorithms/merge_sort_with_inversion_count.py
@@ -1,9 +1,8 @@
-import math
 from typing import Tuple
 def merge_sort_with_inversion_count(arr: list, start: int, end: int) -> Tuple[list, int]:
     inversion_count = 0
     if start < end:
-        mid = start + math.floor((end - start) / 2)
+        mid = (start + end) // 2
         # Sort left array
         arr, inversion_count_left = merge_sort_with_inversion_count(arr, start, mid)
         # Sort right array
@@ -14,49 +13,38 @@ def merge_sort_with_inversion_count(arr: list, start: int, end: int) -> Tuple[li
     
     return arr, inversion_count
 
-def merge(arr: list, start: int, mid: int, end: int) -> Tuple[list, int]:
+def merge(arr: list, start: int, mid: int, end: int) -> int:
     inversion_count = 0
-    n_left = mid + 1 - start
-    n_right = end - mid
-    
-    arr_left = []
-    arr_right = []
-    
-    # Copy all relevant elements from arr to arr_left and arr_right
-    for i in range(0, n_left): 
-        arr_left.append(arr[start + i])
 
-    for j in range(0, n_right): 
-        arr_right.append(arr[mid + 1 + j])
-        
-    i = 0
-    j = 0
+    left = arr[start:mid + 1]
+    right = arr[mid + 1:end + 1]
+    
+    i = j = 0
     k = start
     
     # Compare one by one the earliest element in the left array (sorted) with the other in the right array (sorted)
-    while i < n_left and j < n_right:
-        if arr_left[i] <= arr_right[j]:
-            arr[k] = arr_left[i]
+    while i < len(left) and j < len(right):
+        if left[i] <= right[j]:
+            arr[k] = left[i]
             i += 1
         else: 
-            arr[k] = arr_right[j]
-            inversion_count += (n_left - i)
+            arr[k] = right[j]
+            inversion_count += (len(left) - i)
             j += 1
         k += 1
         
-    # If the left array is exhausted, put all remaining elements in the right array to the remaining slots
-    if i == n_left:
-        while j < n_right: 
-            arr[k] = arr_right[j]
-            j += 1
-            k += 1
-    # If the right array is exhausted, put all remaining elements in the left array to the remaining slots
-    # Also increment inversion_count by the number of right elements, since the single left element is smaller than all elements on the right
-    if j == n_right:
-        while i < n_left:
-            arr[k] = arr_left[i]
-            i += 1
-            k += 1
+    # If the left array is not yet exhausted, put all remaining elements in the left array to the remaining slots
+    while i < len(left):
+        arr[k] = left[i]
+        i += 1
+        k += 1
+            
+    # If the left array is not yet exhausted, put all remaining elements in the right array to the remaining slots
+    while j < len(right):
+        arr[k] = right[j]
+        j += 1
+        k += 1
+        
     return arr, inversion_count
 
 if __name__ == '__main__':